Refactor moving config into MctPlot component. Fix Legend issues.
